pyproject.toml
[project]
name = "water-analysis"
version = "1.0.0"
description = "水质三维智能监测与分析系统"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "sqlalchemy[asyncio]>=2.0.30",
    "aiomysql>=0.2.0",
    "pymysql>=1.1.1",
    "redis>=5.0.6",
    "celery>=5.4.0",
    "pydantic>=2.7.3",
    "pydantic-settings>=2.3.3",
    "python-multipart>=0.0.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.1",
    "scipy>=1.14.0",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.4",
    "plotly>=5.22.0",
    "pandas>=2.2.2",
    "kaleido>=0.2.1",
    "python-docx>=1.1.2",
    "openai>=1.35.1",
    "pymilvus>=2.4.4",
    "minio>=7.2.7",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py311"
